+++ assets/voxygen/shaders/include/sky.glsl
@@ -26,7 +26,7 @@ const vec3 DAWN_LIGHT   = vec3(5.0, 2.0, 1.15);
 const vec3 SUN_HALO_DAWN = vec3(8.2, 3.0, 2.1);
 
 const vec3 SKY_DAY_TOP = vec3(0.1, 0.5, 0.9);
-const vec3 SKY_DAY_MID = vec3(0.02, 0.28, 0.8);
+const vec3 SKY_DAY_MID = vec3(0.18, 0.28, 0.6);
 const vec3 SKY_DAY_BOT = vec3(0.1, 0.2, 0.3);
 const vec3 DAY_LIGHT   = vec3(3.8, 3.0, 1.8);
 const vec3 SUN_HALO_DAY = vec3(0.25, 0.25, 0.001);
@@ -503,7 +503,12 @@ float is_star_at(vec3 dir) {
 
     //return 0.0;
 
-    return 5.0 / (1.0 + pow(dist * 750, 8));
+    #if (CLOUD_MODE == CLOUD_MODE_NONE)
+        const float power = 5.0;
+    #else
+        const float power = 50.0;
+    #endif
+    return power * max(sun_dir.z, 0.1) / (1.0 + pow(dist * 750, 8));
 }
 
 vec3 get_sky_light(vec3 dir, float time_of_day, bool with_stars) {
@@ -570,7 +575,7 @@ vec3 get_sky_light(vec3 dir, float time_of_day, bool with_stars) {
     return sky_color * magnetosphere_tint;
 }
 
-vec3 get_sky_color(vec3 dir, float time_of_day, vec3 origin, vec3 f_pos, float quality, bool with_features, float refractionIndex) {
+vec3 get_sky_color(vec3 dir, float time_of_day, vec3 origin, vec3 f_pos, float quality, bool with_features, float refractionIndex, bool fake_clouds, float sun_shade_frac) {
     // Sky color
     /* vec3 sun_dir = get_sun_dir(time_of_day);
     vec3 moon_dir = get_moon_dir(time_of_day); */
@@ -581,7 +586,7 @@ vec3 get_sky_color(vec3 dir, float time_of_day, vec3 origin, vec3 f_pos, float q
     // moon_dir = moon_dir.z <= 0 ? refract(moon_dir/*-view_dir*/, vec3(0.0, 0.0, 1.0), refractionIndex) : moon_dir;
 
     // Sun
-    const vec3 SUN_SURF_COLOR = vec3(1.5, 0.9, 0.35) * 50.0;
+    const vec3 SUN_SURF_COLOR = vec3(1.5, 0.9, 0.35) * 10.0;
 
     vec3 sun_halo_color = mix(
         (sun_dir.x > 0 ? SUN_HALO_DUSK : SUN_HALO_DAWN)* magnetosphere_tint,
@@ -591,15 +596,29 @@ vec3 get_sky_color(vec3 dir, float time_of_day, vec3 origin, vec3 f_pos, float q
 
     float sun_halo_power = 20.0;
     #if (CLOUD_MODE == CLOUD_MODE_NONE)
-        sun_halo_power = 1000.0;
-        sun_halo_color *= 0.1;
+        if (true) {
+    #else
+        if (fake_clouds || medium.x == MEDIUM_WATER) {
     #endif
+        sun_halo_power = 30.0;
+        sun_halo_color *= 0.01;
+    }
 
     vec3 sun_halo = sun_halo_color * 25 * pow(max(dot(dir, -sun_dir), 0), sun_halo_power);
     vec3 sun_surf = vec3(0);
     if (with_features) {
         float angle = 0.00035;
-        sun_surf = clamp((dot(dir, -sun_dir) - (1.0 - angle)) * 4 / angle, 0, 1) * SUN_SURF_COLOR * SUN_COLOR_FACTOR;
+        sun_surf = clamp((dot(dir, -sun_dir) - (1.0 - angle)) * 4 / angle, 0, 1)
+            * SUN_SURF_COLOR
+            * SUN_COLOR_FACTOR
+            * sun_shade_frac;
+    }
+    #if (CLOUD_MODE == CLOUD_MODE_NONE)
+        if (true) {
+    #else
+        if (fake_clouds || medium.x == MEDIUM_WATER) {
+    #endif
+        sun_surf *= 0.1;
     }
     vec3 sun_light = sun_halo + sun_surf;
 
@@ -610,25 +629,34 @@ vec3 get_sky_color(vec3 dir, float time_of_day, vec3 origin, vec3 f_pos, float q
     vec3 moon_halo_color = MOON_HALO_COLOR;
 
     float moon_halo_power = 20.0;
-    #if (CLOUD_MODE == CLOUD_MODE_NONE)
-        moon_halo_power = 2500.0;
-        moon_halo_color *= 0.1;
-    #endif
 
-    vec3 moon_halo = moon_halo_color * pow(max(dot(dir, -moon_dir), 0), moon_halo_power);
     vec3 moon_surf = vec3(0);
     if (with_features) {
         float angle = 0.00035;
         moon_surf = clamp((dot(dir, -moon_dir) - (1.0 - angle)) * 4 / angle, 0, 1) * MOON_SURF_COLOR;
     }
+    #if (CLOUD_MODE == CLOUD_MODE_NONE)
+        if (true) {
+    #else
+        if (fake_clouds || medium.x == MEDIUM_WATER) {
+    #endif
+        moon_halo_power = 50.0;
+        moon_halo_color *= 0.2;
+        moon_surf *= 0.05;
+    }
+    vec3 moon_halo = moon_halo_color * pow(max(dot(dir, -moon_dir), 0), moon_halo_power);
     vec3 moon_light = moon_halo + moon_surf;
 
     // Replaced all clamp(sun_dir, 0, 1) with max(sun_dir, 0) because sun_dir is calculated from sin and cos, which are never > 1
 
+    vec3 sky_color;
     #if (CLOUD_MODE == CLOUD_MODE_NONE)
-        vec3 sky_color = get_sky_light(dir, time_of_day, true);
+        if (true) {
     #else
-        vec3 sky_color;
+        if (fake_clouds || medium.x == MEDIUM_WATER) {
+    #endif
+        sky_color = get_sky_light(dir, time_of_day, !fake_clouds);
+    } else {
         if (medium.x == MEDIUM_WATER) {
             sky_color = get_sky_light(dir, time_of_day, true);
         } else {
@@ -636,13 +664,17 @@ vec3 get_sky_color(vec3 dir, float time_of_day, vec3 origin, vec3 f_pos, float q
             float star = is_star_at(star_dir);
             sky_color = vec3(0) + star;
         }
-    #endif
+    }
 
     return sky_color + sun_light + moon_light;
 }
 
+vec3 get_sky_color(vec3 dir, float time_of_day, vec3 origin, vec3 f_pos, float quality, bool with_features, float refractionIndex) {
+    return get_sky_color(dir, time_of_day, origin, f_pos, quality, with_features, refractionIndex, false, 1.0);
+}
+
 vec3 get_sky_color(vec3 dir, float time_of_day, vec3 origin, vec3 f_pos, float quality, bool with_stars) {
-    return get_sky_color(dir, time_of_day, origin, f_pos, quality, with_stars, 1.0);
+    return get_sky_color(dir, time_of_day, origin, f_pos, quality, with_stars, 1.0, false, 1.0);
 }
 
 float fog(vec3 f_pos, vec3 focus_pos, uint medium) {